+++ NEWS.md
@@ -17,6 +17,7 @@
 - Added: `as.polygons()` `data.frame` method for correctly formatted data.tables (replaces internal `dt_to_spatvector_points()`)
 - Added: autocompletes for `$` with `spatLocsObj`, `spatEnrObj`, `dimObj`, `cellMetaObj`, `featMetaObj`, `giottoPolygon`, `giottoPoints`
 - Added: `toplevel_params` param to `subsetGiottoLocs()`
+- Added: `spin`, `spatShift`, `rescale` methods for `data.frame`
 
 ## Changes
 - Fixed: Provenance now not accidentally created as a list in `addSpatialCentroidLocations()`